Use a more robust method to avoid import renames in `$imports` init
Depending on how Babel transforms are applied, it is possible for the
`ImportMap` import to end up being used as `var $imports =
helpers.ImportMap` which breaks the previous check to avoid replacing
the reference to the import in the arguments passed to the `ImportMap`
constructor.

Resolve this by avoiding any replacements when the identifier is a child
AST node of a variable declaration for a variable named `$imports`.