Register imports after their declaration instead of all at once
ES6 imports are hoisted so that all declarations should occur before any
code that refers to them. With CommonJS however this is not the case, so
we may have:

```
var moduleA = require("module-a");
var ident = moduleA.ident;
var moduleB = require("module-b");
```

In this case we need to rewrite the `ident` initialization to
`$imports.moduleA.ident`, which requires that `$imports` is defined and
the `moduleA` import is registered, before the "module-b" import has
been processed.

Solve this by registering imports one at a time using `$imports.$add`
method calls instead of registering them in the `ImportMap` constructor
after all of the import/require statements in the file.