Transform variable declarations before object patterns are transformed
Variable declarations with object patterns, such as:

```
var { symbol } = require("a-module")
```

Are transformed by Babel into ES5 as:

```
var _aModule = require("a-module"), symbol = _aModule.symbol
```

If this transform happened before the plugin visits the `require(...)`
call expression, the `symbol` variable was not marked as an import and
references to it were not transformed into `$imports.symbol`.  Therefore
it was not possible to mock them.

Fix this problem by visiting the `VariableDeclaration` node in the main
visitor instead of the `CallExpression`. Provided that the
mockable-imports plugin visits the node before the object pattern
transform, the plugin will be able to see the `symbol` identifier and
register it as an import.