Make generated `$imports.$add(...)` calls statements
Previously the generated `$imports.$add` calls were generated as just
`CallExpression`s instead of statements. Top-level items in a module
should all be statements or declarations per the ECMAScript spec. I'm
not entirely sure how Babel handles invalid ASTs but this caused
problems when the plugin was used together with
babel-plugin-transform-async-to-promises.