Handle identifier for import being renamed by another plugin
Handle the case where an identifier referring to an import is renamed by
another Babel plugin after the `$imports.$add` call is generated for
that identifier but before the identifier is referenced in some later
context.

For example, given:

```
import { render } from "preact";

var x = {
  render: function() {
    render();
  }
};
```

The @babel/plugin-transform-function-name plugin, if active, would transform the
code to:

```
import { render as _render } from "preact";

var x = {
  render: function() {
    _render();
  }
};
```

When processing the call to `_render()` we need to generate a call to
`$imports.render()` rather than `$imports._render()` because earlier in
the module we will have generated a call to `$imports.$add("render",
"preact", "render", render)` where the first argument is the alias name.

The solution implemented here is to record the alias name used when
inserting the `$imports.$add` call and to use that alias name instead of
the identifier's current name when replacing later references to the
identifier.

Fixes #24