Exclude Rollup helper modules from mocking
@rollup/plugin-babel may add an import of `\0rollupPluginHelpers.js` at
the top of a module in a way that can lead to usage of the `$imports`
helper generated by this plugin before its declaration. Since there is
no value in mocking this import, this commit resolves the issue by
adding these helper modules to the plugin's exclusion list. In the
process the `excludeImportsFromModules` option's syntax was expanded to
support regex patterns rather than just strings.