Drop CommonJS support
All of the downstream projects which are the primary consumers of this plugin
switched to ES imports some years ago.

Make future maintenance easier by dropping support for transforming CommonJS
modules. Note that this only affects modules whose imports are mocked by this
plugin. If such a module imports a third-party module written in CommonJS, that
will still work.